Alignment and HMM files
Alignment of Spike proteins to identify highly conserved residues to be used as numbering bases. Spike_PDB.hmm was used to identify the proteins used for Swiss_Prot_Spikes.hmm and the alignment Spikes.aln includes both sets. I have been focusing on the sections that align to 60MJ_E which is specifically the receptor binding domain.